refactor: [#76] extract stock module
