+++ src/Domain/@shared/Factory/RepositoryFactory.ts
@@ -1,7 +1,7 @@
 import { CouponRepository } from "../../Checkout/Repository/CouponRepository";
 import { OrderRepository } from "../../Checkout/Repository/OrderRepository";
 import { ProductRepository } from "../../Product/Repository/ProductRepository";
-import { StockEntryRepository } from "../../Stock/Repository/StockEntryRepository";
+import { StockEntryRepository } from "../../../modules/Stock/Domain/Repository/StockEntryRepository";
 
 export interface RepositoryFactory {
   makeOrderRepository(): OrderRepository;
