+++ src/modules/Stock/Controller/StockController.ts
@@ -1,7 +1,7 @@
-import { StockHandler } from "../../../Application/Stock/handler/StockHandler";
+import { StockHandler } from "../../../modules/Stock/Handler/StockHandler";
 import { RepositoryFactory } from "../../../Domain/@shared/Factory/RepositoryFactory";
 import { OrderPlaced } from "../../../Domain/Checkout/Event/OrderPlaced";
-import { Queue } from "../../@shared/Queue/Queue";
+import { Queue } from "../../../Infra/@shared/Queue/Queue";
 
 export class StockController {
   constructor(queue: Queue, repositoryFactory: RepositoryFactory) {
