+++ src/modules/Stock/Domain/Repository/StockEntryRepository.ts
@@ -1,5 +1,5 @@
 import { StockEntry } from "../Entity/StockEntry";
-import { Id } from "../../@shared/ValueObject/Id";
+import { Id } from "../../../../Domain/@shared/ValueObject/Id";
 
 export interface StockEntryRepository {
   save(stockEntry: StockEntry): Promise<void>;
