+++ src/modules/Stock/Domain/Service/StockCalculator.ts
